Implement the bulk of the group-events Lamdba which will help sequence writes
This approach should help address some problems identified in [this blog
post](https://www.buoyantdata.com/blog/2023-11-27-concurrency-limitations-with-deltalake-on-aws.html).
In real-world scenarios lock acquisition timeouts will happen if a large sync
results in a substantial number of parquet files being dropped into the same S3
table prefix.

The simple oxbow deployment is:

    S3 Events -> SQS -> oxbow

This approach sequences events into a FIFO queue which should help avoid lock
contention:

    S3 Events -> SQS -> group-events -> SQS FIFO -> oxbow

The use of the table prefix as the message group ID ensures that the oxbow
lambda will not be invoked concurrently for the table prefix