Introduce significant changes to the file-loader Lambda to handle larger inputs
The original design of the file-loader lambda was oriented around
streams of _small_ files landing in S3. With a recent use-case a single
file-loader Lambda needed to alternate between handling many small files
and many _large_ files.

The previous mode of operation led to out of memory errors as the lambda
tried to read more data than it should from S3 before writing.

With this mode, the batch size should always be 1 for the file-loader
lambda's trigger configuration, but then it should be configured to
retrieve additional messages from SQS similar to the "BUFFER_MORE"
configuration with SQS.

This change also refactors the "BUFFER MORE" functionality out of
sqs-ingest into the oxbow-sqs crate