Properly iterate through record batches on deserialization
Testing this in an automated fashion proved to be majority of the
challenge since the aws-sdk-s3 crate can produce an AsyncBufRead
implementation for asynchronously reading through the S3 input file, but
chaining that together within an abstracted function for execution.

This change also includes the introduction of stats_alloc to try to do a
better job tracking memory as files are loaded.