Turn off expired logs cleanup
When setting is enabled (by default), it scans delta logs on each
commit, and that leads to a significant overhead.

It's possible to turn it off on table level, but that causes logs never
deleted (even on vacuum). By turning the setting off on commit, the logs
are still cleaned up when table is vacuumed or optimized, but not when
data is added